% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State_Neighbors.R
\docType{data}
\name{State_Neighbors}
\alias{State_Neighbors}
\title{US State Neighbor List}
\format{
A data frame with 50 observations and 2 variables.
\tabular{ll}{ \cr
 state \tab state name \cr
 neighbors \tab character string of neighboring states (separated by ',') for each state observation \cr
}
}
\usage{
State_Neighbors
}
\description{
This data set provides a list (as a character string) of neighboring states for each U.S. state.
}
\keyword{datasets}
