% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_Counties.R
\docType{data}
\name{US_Counties}
\alias{US_Counties}
\title{US Counties Information for Merging}
\format{
A data frame with 3106 observations and 11 variables.
\tabular{ll}{ \cr
 countystate \tab proper county name and state name (separated by ':') \cr
 state_name \tab proper state name \cr
 county_name \tab proper county name \cr
 state_abbv \tab abbreviated state name \cr
state_name_cap \tab capitalized state name \cr
state_name_cap_nominate \tab capitalized state name, shortened (as in DW-NOMINATE data) \cr
state_fips \tab state FIPS \cr
county_fips \tab county FIPS \cr
fullgeoid \tab Twelve-digit GEOID identifier for each county observation (as a character string) \cr
neighbors_fullgeoid \tab character string of neighboring counties' fullgeoid's (separated by ',') for each county observation \cr
neighbors_countystate \tab character string of neighboring counties' countystate names (separated by ',') for each county observation \cr
}
}
\usage{
US_Counties
}
\description{
This data set provides common names and abbreviations for U.S. counties to enable merging with various data sets.
}
\keyword{datasets}
